Moved interface structs to their own module
